+++ src/cmd/transfer.rs
@@ -1,4 +1,4 @@
-use crate::ad_hoc_route;
+use super::runtime::ad_hoc_route;
 use binport::progress::TransferProgress;
 use binport::remote_command;
 use binport::ssh::{Destination, NativeSsh};